TEST: Renaming base endpoint to explore and updated folder structure for scanners test
