+++ tests/io/v3/vm/scanners/test_schema.py
